Cast to double for double division

+++ week-3-pattern-recognition/src/Point.java
@@ -69,7 +69,7 @@ public class Point implements Comparable<Point> {
             return Double.POSITIVE_INFINITY;
         }
 
-        return (that.y - this.y) / (that.x - this.x);
+        return (that.y - this.y) / (double)(that.x - this.x);
     }
 
     /**
